pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "tiledpy"
version = "0.1.0"
description = "TMX map loader using Pillow for sprite detection and Pygame for rendering"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pillow>=10.0.0",
    "pygame>=2.5.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]
docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "pymdown-extensions>=10.0",
    "mkdocstrings[python]>=0.25.0",
]

[tool.hatch.build.targets.wheel]
packages = ["tiledpy"]
